+++ web/src/vite-env.d.ts
@@ -2,3 +2,18 @@ declare module '*.svg' {
   const url: string
   export default url
 }
+
+declare module '*.png' {
+  const url: string
+  export default url
+}
+
+declare module '*.jpeg' {
+  const url: string
+  export default url
+}
+
+declare module '*.jpg' {
+  const url: string
+  export default url
+}